refactor(Metrics): :recycle: Add metric utils
